Moving quote and joke code to their own folder and fixing/making delete_quote better
